✨ convert torch model to tflite
Convert pytorch model to tflite model and test the tflite model by
feeding test images as inut data and visualize the output results.
During training model weights are save in `.pt` format. It cannot be
used directly on portable devices. To reduce the size of the model
and make the trained results ipad friendly, convert torch format
weights into `tflite` models. The output of the tflite model is not
human readable thus post process and visualize them on the
corresponding input images for testing.

Co-authored-by: Mert Sayar <mert.sayar@gmail.com>